+++ Library/Aphid/UI/Controls.js
@@ -5,3 +5,4 @@
 Aphid.UI.Controls = {};
 
 //= require "Controls/SelectView"
+//= require "Controls/DateView"